Consolidate dask and mlflow fixtures
