+++ tests/examples/ransomware_detection/conftest.py
@@ -20,7 +20,6 @@ import pytest
 import yaml
 
 from _utils import TEST_DIRS
-from _utils import import_or_skip
 from _utils import remove_module
 
 # pylint: disable=redefined-outer-name
@@ -30,12 +29,12 @@ SKIP_REASON = ("Tests for the ransomware_detection example require a number of p
                "for details on installing these additional dependencies")
 
 
-@pytest.fixture(autouse=True, scope='session')
-def dask_distributed(fail_missing: bool):
+@pytest.fixture(name="dask_distributed", autouse=True, scope='session')
+def dask_distributed_fixture(dask_distributed):
     """
     All of the tests in this subdir requires dask.distributed
     """
-    yield import_or_skip("dask.distributed", reason=SKIP_REASON, fail_missing=fail_missing)
+    yield dask_distributed
 
 
 @pytest.fixture(name="config")
